backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mistralai
elevenlabs
python-multipart
